+++ CHANGELOG.md
@@ -8,6 +8,12 @@ We attempt to adhere to [Semantic Versioning](http://semver.org/).
 ### Removed
 ### Changed
 
+## [1.5.2] - 2017-03-29
+### Added
+- User can map rbd image through rbd-nbd command now
+### Changed 
+- fix ServeUnix() gid parameter problem
+
 ## [1.5.1] - 2016-09-20
 ### Changed
 - Update XFS checks to mount/unmount to clear any disk logs